SCGCN-main/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_tflow_select=2.1.0=gpu
absl-py=0.8.1=py27_0
astor=0.8.0=py27_0
backports=1.0=py_2
backports.weakref=1.0.post1=py_1
blas=1.0=mkl
c-ares=1.15.0=h7b6447c_1001
ca-certificates=2020.1.1=0
certifi=2019.11.28=py27_0
cffi=1.13.2=py27h2e261b9_0
cudatoolkit=10.0.130=0
cudnn=7.6.5=cuda10.0_0
cupti=10.0.130=0
enum34=1.1.6=py27_1
freetype=2.9.1=h8a8886c_1
funcsigs=1.0.2=py27_0
future=0.18.2=py27_0
futures=3.3.0=py27_0
gast=0.2.2=py27_0
google-pasta=0.2.0=py_0
grpcio=1.27.2=py27hf8bcb03_0
h5py=2.9.0=py27h7918eee_0
hdf5=1.10.4=hb1b8bf9_0
intel-openmp=2020.0=166
jpeg=9b=h024ee3a_2
keras-applications=1.0.8=py_0
keras-preprocessing=1.1.0=py_1
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.11.2=hd408876_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_0
linecache2=1.0.0=py27_0
markdown=3.1.1=py27_0
mkl=2020.0=166
mkl-service=2.3.0=py27he904b0f_0
mkl_fft=1.0.15=py27ha843d7b_0
mkl_random=1.1.0=py27hd6b4f25_0
mock=2.0.0=py27_0
ncurses=6.2=he6710b0_1
ninja=1.9.0=py27hfd86e86_0
numpy=1.16.6=py27hbc911f0_0
numpy-base=1.16.6=py27hde5b4d6_0
olefile=0.46=py27_0
openssl=1.1.1g=h7b6447c_0
opt_einsum=2.3.2=py_0
pbr=5.4.4=py_0
pillow=6.2.1=py27h34e0f95_0
pip=19.3.1=py27_0
protobuf=3.11.2=py27he6710b0_0
pycparser=2.20=py_0
python=2.7.18=h02575d3_0
pytorch=1.1.0=py2.7_cuda10.0.130_cudnn7.5.1_0
readline=8.0=h7b6447c_0
scipy=1.2.1=py27h7c811a0_0
setuptools=44.0.0=py27_0
six=1.13.0=py27_0
sqlite=3.31.1=h62c20be_1
tensorboard=2.0.0=py27hb38c66f_1
tensorflow=2.0.0=gpu_py27hb041a2f_0
tensorflow-base=2.0.0=gpu_py27h356bb79_0
tensorflow-estimator=2.0.0=pyh2649769_0
tensorflow-gpu=2.0.0=h0d30ee6_0
termcolor=1.1.0=py27_1
tk=8.6.8=hbc83047_0
torchvision=0.3.0=py27_cu10.0.130_1
traceback2=1.4.0=py27_0
unittest2=1.1.0=py27_0
werkzeug=0.16.1=py_0
wheel=0.33.6=py27_0
wrapt=1.11.2=py27h7b6447c_0
xz=5.2.5=h7b6447c_0
zlib=1.2.11=h7b6447c_3
zstd=1.3.7=h0b5b093_0
